+++ stroom-pipeline/src/main/java/stroom/pipeline/server/filter/SourceLocationFactory.java
@@ -1,0 +1,7 @@
+package stroom.pipeline.server.filter;
+
+public interface SourceLocationFactory {
+    void reset();
+
+    void storeLocation();
+}